Remove Option from Filesystem::files() iterator
The iterator returned from Filesystem::files() should yield Entry,
rather than Option<Entry>. If a filesystem entry does not contain a
file, this iterator should skip forward until it finds a valid file
instead of yielding a None.

An alternative would be to yield (None, index), but yielding None alone
is not useful to most users and only adds to confusion.